backend/requirements.txt
# backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
cryptography>=42.0.0
python-dotenv>=1.0.0
aiofiles>=23.2.0

# 数据库驱动
pgvector>=0.2.0
psycopg2-binary>=2.9.9
clickhouse-connect>=0.6.0
mysql-connector-python>=8.0.0
redis>=5.0.0
elasticsearch>=8.12.0
minio>=7.2.0

# 导出处理
openpyxl>=3.1.0
xlsxwriter>=3.1.9

# AI 相关
httpx>=0.26.0

# SQL 处理
sqlparse>=0.4.0